requirements.txt
# List of dependencies for Smart Music Player
mutagen
pydub
numpy
scikit-learn
tk
librosa
pillow
pygame